requirements/base.txt
# Core Django packages (Python 3.14 compatible)
Django==5.0.6
djangorestframework>=3.16.0
django-cors-headers==4.3.1

# Database
psycopg2-binary>=2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0

# Environment variables
python-decouple==3.8

# Redis and Celery
redis==5.0.1
celery==5.3.4
django-redis==5.4.0

# WebSocket support
channels[daphne]==4.0.0
channels-redis==4.1.0

# SMS & Push Notifications
twilio==8.10.0
pywebpush==1.14.0
py-vapid==1.9.0

# Image processing
Pillow>=10.3.0

# API documentation
drf-spectacular==0.27.0

# Admin UI
django-unfold>=0.75.0

# Static files serving for production
whitenoise>=6.6.0

# Web server
gunicorn>=21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
